Move header related stuff into subdir
